third_party/WebKit/LayoutTests/xr/navigator_xr_blocked_by_getVRDisplays.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

promise_test((t) => navigator.getVRDisplays().then(() => {
  assert_true(navigator.xr == null);
}), "Test that access to navigator.xr is blocked once navigator.getVRDisplays has been called.");

</script>
